timeout-bandaid
Due to a bug in the archive-upload service, it actively handles only one
request at a time, whilst accepting several and queueing them. This
means that the slowest of the API-endpoints (uploading) can hold up any
of the other endpoints requested from other archive-tasks performed in
parallel. Stackstorm will mark these tasks (and subsequently the
workflow) as failed, but the archive-upload service will actually
perform them once it gets around to doing so.
Until this is fixed, we increase the timeout on all tasks performed at
the archive-upload service.